+++ CHANGELOG.md
@@ -132,6 +132,7 @@ The following plugins were added:
 - Regex: Search()
 - Regex: Replace()
 - Rename: SetPCNameOverride()
+- Rename: GetPCNameOverride()
 - Reveal: RevealTo()
 - Reveal: SetRevealToParty()
 - Util: GenerateUUID()
@@ -165,6 +166,7 @@ The following plugins were added:
 ### Fixed
 - Administration: Fixed a nullptr crash in Get{DM/Player}Password()
 - Chat: Fix recursive message skipping
+- Object: GetLocalVariable and GetLocalVariableCount now work with Area and Module object
 - Events: DMActionEvents now return valid TARGET event data
 
 [Unreleased]: https://github.com/nwnxee/unified/compare/build8186-2...HEAD